backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
pdfplumber
python-docx
PyPDF2
google-generativeai
python-dotenv
fpdf2
gunicorn
